lambda/ai_processor/requirements.txt
google-generativeai==0.3.2
urllib3==1.26.18
